add also an ubuntu-based build: this is needed because the centos6 wheel are not usable on ubuntu, until auditwheel works

+++ docker/image/install_packages.sh
@@ -1,4 +1,16 @@
 #!/bin/bash
 
-# packages needed to build numpy and scipy
-yum install -y atlas-devel blas-devel lapack-devel
+if [ -f /etc/debian_version ]
+then
+    # we are on ubuntu
+    export DEBIAN_FRONTEND noninteractive
+    apt-get update
+    apt-get install -y curl gcc g++
+
+    # needed for numpy and scipy
+    apt-get install gfortran libatlas-dev libblas-dev liblapack-dev
+else
+    # we are on centos
+    # packages needed to build numpy and scipy
+    yum install -y atlas-devel blas-devel lapack-devel
+fi
